flask-server/requirements.txt
flask
mysql-connector-python
paho-mqtt
flask-session
redis
python-dotenv
flask-cors